+++ Project-Kothin/Ticket.cs
@@ -16,5 +16,36 @@ namespace Project_Kothin
         {
             InitializeComponent();
         }
+
+        private void label1_Click(object sender, EventArgs e)
+        {
+
+        }
+
+        private void comboBox2_SelectedIndexChanged(object sender, EventArgs e)
+        {
+
+        }
+
+        private void comboBox1_SelectedIndexChanged(object sender, EventArgs e)
+        {
+
+        }
+
+        private void label2_Click(object sender, EventArgs e)
+        {
+
+        }
+
+        private void SEARCHTRAIN_Click(object sender, EventArgs e)
+        {
+            Selectseats s1 = new Selectseats();
+            s1.Show();
+        }
+
+        private void Train_Load(object sender, EventArgs e)
+        {
+
+        }
     }
 }
